Services/Ai.API/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
tensorflow==2.18.0
scikit-learn==1.5.2
numpy==2.0.2
aio-pika==9.4.3
python-dotenv==1.0.1
pydantic==2.9.2
